+++ src/web/assets/ckeditor/src/footnote.js
@@ -1,0 +1,15 @@
+import { Plugin } from "ckeditor5/src/core.js";
+import FootnoteEditing from "./footnoteediting.js";
+import FootnoteUI from "./footnoteui.js";
+
+import "../theme/footnote.css";
+
+export default class Footnote extends Plugin {
+  static get pluginName() {
+    return "Footnote";
+  }
+
+  static get requires() {
+    return [FootnoteEditing, FootnoteUI];
+  }
+}