+++ src/web/assets/ckeditor/src/index.js
@@ -1,0 +1,1 @@
+export { default as Footnote } from "./footnote.js";